refactor(llmproxy): address abstraction violations in inline-approval path
PR self-review surfaced three lower-package layering violations
introduced by the AskUserQuestion work. All addressed in-place so
this PR ships at a clean abstraction boundary.

1. AvailableTools off ApprovalContext.
   AvailableTools is per-request request metadata, not an approval-
   flow dependency. Moved to PostprocessConfig's top level, where
   any stage that needs to branch on what the harness declared can
   read it without expanding ApprovalContext's charter.

2. historystrip stays harness-tool-name-agnostic.
   The synthetic-tool-use orphan finder was hardcoded against
   `name == "AskUserQuestion"`. Replaced with a prefix-match
   against SyntheticToolUseIDPrefix (`toolu_clawvisor_`), the
   namespace the intercept already stamps on every synthesized
   tool_use_id. The constant lives in historystrip and is re-
   exported from llmproxy so producer/consumer stay in lockstep.
   historystrip no longer knows about any specific harness tool.

3. AskUserQuestion parsing out of conversation, deduped.
   conversation/approval_release.go was hardcoding "AskUserQuestion"
   and carrying ~150 lines of AskUserQuestion-shape detection plus
   helpers. Moved the entire AskUserQuestion-aware branch (parser,
   wrapper-unwrap, sibling-text scoping, tool_result extraction)
   into a new llmproxy/askuserquestion_reply.go. The body editor's
   duplicate collectAnthropicAskUserQuestionInputs / tool_result
   helpers are gone — both detector and rewriter share one
   findAnthropicAskUserQuestionMatch finder so they can never
   drift on which tool_result counts as the answer.

   anthropicApprovalBodyEditor.LatestApprovalReply now tries the
   text path via conversation.AnthropicApprovalReply, then falls
   back to anthropicAskUserQuestionApprovalReply in llmproxy. The
   conversation package is back to wire shapes + generic Clawvisor-
   marker conventions, with no harness-tool-name knowledge.

Tests for the AskUserQuestion shape moved alongside the code into
internal/runtime/llmproxy/askuserquestion_reply_test.go.
