+++ internal/api/handlers/llm_endpoint.go
@@ -1181,7 +1181,6 @@ func (h *LLMEndpointHandler) serve(w http.ResponseWriter, r *http.Request) {
 					InlineTaskCreator:                h.InlineTaskCreator,
 					Checkouts:                        h.TaskCheckouts,
 					DefaultTaskExpirySeconds:         h.DefaultTaskExpirySeconds,
-					AvailableTools:                   reqSummary.AvailableTools,
 				},
 				RewriteContext: llmproxy.RewriteContext{
 					Inspector:    h.Inspector,
@@ -1195,6 +1194,7 @@ func (h *LLMEndpointHandler) serve(w http.ResponseWriter, r *http.Request) {
 				RoutingContext: llmproxy.RoutingContext{
 					ControlBaseURL: h.ControlBaseURL,
 				},
+				AvailableTools: reqSummary.AvailableTools,
 			}
 			// First-turn routing notice. Mirrors the buffered path's
 			// invocation below (search "first-turn routing notice").
@@ -1591,7 +1591,6 @@ func (h *LLMEndpointHandler) serve(w http.ResponseWriter, r *http.Request) {
 				InlineTaskCreator:                h.InlineTaskCreator,
 				Checkouts:                        h.TaskCheckouts,
 				DefaultTaskExpirySeconds:         h.DefaultTaskExpirySeconds,
-				AvailableTools:                   reqSummary.AvailableTools,
 			},
 			RewriteContext: llmproxy.RewriteContext{
 				Inspector:    h.Inspector,
@@ -1605,6 +1604,7 @@ func (h *LLMEndpointHandler) serve(w http.ResponseWriter, r *http.Request) {
 			RoutingContext: llmproxy.RoutingContext{
 				ControlBaseURL: h.ControlBaseURL,
 			},
+			AvailableTools: reqSummary.AvailableTools,
 		})
 		h.Logger.DebugContext(r.Context(), "lite-proxy postprocess complete",
 			"request_id", requestID,
@@ -1662,7 +1662,6 @@ func (h *LLMEndpointHandler) serve(w http.ResponseWriter, r *http.Request) {
 					InlineTaskCreator:                h.InlineTaskCreator,
 					Checkouts:                        h.TaskCheckouts,
 					DefaultTaskExpirySeconds:         h.DefaultTaskExpirySeconds,
-					AvailableTools:                   reqSummary.AvailableTools,
 				},
 				RewriteContext: llmproxy.RewriteContext{
 					Inspector:    h.Inspector,
@@ -1676,6 +1675,7 @@ func (h *LLMEndpointHandler) serve(w http.ResponseWriter, r *http.Request) {
 				RoutingContext: llmproxy.RoutingContext{
 					ControlBaseURL: h.ControlBaseURL,
 				},
+				AvailableTools: reqSummary.AvailableTools,
 			})
 			switch {
 			case contErr != nil:
