+++ internal/runtime/llmproxy/config.go
@@ -141,12 +141,6 @@ type ApprovalContext struct {
 	ConversationAutoApproveThreshold string
 	Checkouts                        TaskCheckoutStore
 	DefaultTaskExpirySeconds         int
-	// AvailableTools is the inbound request's declared tool list
-	// (e.g. Anthropic tools[].name). The inline approval renderer
-	// reads it to decide whether to instruct the model to surface
-	// the yes/no via AskUserQuestion (when the harness exposes that
-	// tool) or to relay the prompt as plain text.
-	AvailableTools []string
 }
 
 // RewriteContext groups the credentialed-rewrite path's dependencies:
@@ -198,6 +192,16 @@ type PostprocessConfig struct {
 	RewriteContext
 	ScriptSessionContext
 	RoutingContext
+
+	// AvailableTools is the inbound request's declared tool list
+	// (e.g. Anthropic tools[].name). Per-request request metadata,
+	// not an approval-flow dependency — sits at the top level so
+	// any stage that needs to branch on what the harness declared
+	// can read it without dragging it into a sub-context whose
+	// charter doesn't cover it. Currently consumed by the inline-
+	// approval intercept (to decide whether to substitute with an
+	// AskUserQuestion picker or a text prompt).
+	AvailableTools []string
 }
 
 // PostprocessResult is what postproc.Postprocess +
